ref(charts): Move STARFISH_FIELDS out of charts.ts (#71634)
This module imports a TON of stuff, STARFISH_FIELDS is used by
discover/fields. By moving this out we can reduce the import graph
generated for chartcuterie by a magnitude of a few MBs.

This is also needed for GH-70632
